Refactored categoryController

+++ PizzeriaButikenOnline/PizzeriaButikenOnline/Controllers/CategoryController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using PizzeriaButikenOnline.Data;
 using PizzeriaButikenOnline.Models;
@@ -10,18 +11,12 @@ namespace PizzeriaButikenOnline.Controllers
     public class CategoryController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly IMapper _mapper;
 
-        public CategoryController(ApplicationDbContext context)
+        public CategoryController(ApplicationDbContext context, IMapper mapper)
         {
             _context = context;
-        }
-
-        public ActionResult Create()
-        {
-            return PartialView("Forms/CategoryFormPartial", new CategoryFormViewModel
-            {
-                Action = nameof(Create)
-            });
+            _mapper = mapper;
         }
 
         [HttpPost]
@@ -29,36 +24,12 @@ namespace PizzeriaButikenOnline.Controllers
         public ActionResult Create(CategoryFormViewModel viewModel)
         {
             if (!ModelState.IsValid)
-            {
                 return PartialView("Forms/CategoryFormPartial", viewModel);
-            }
 
-            try
-            {
-                _context.Categories.Add(new Category
-                {
-                    Name = viewModel.Name
-                });
-                _context.SaveChanges();
+            _context.Categories.Add(_mapper.Map<CategoryFormViewModel, Category>(viewModel));
+            _context.SaveChanges();
 
-                return RedirectToAction(nameof(Index), "Home");
-            }
-            catch
-            {
-                return PartialView("Forms/CategoryFormPartial", viewModel);
-            }
-        }
-
-        public ActionResult Edit(int id)
-        {
-            var category = _context.Categories.Find(id);
-            var viewModel = new CategoryFormViewModel
-            {
-                Id = category.Id,
-                Name = category.Name,
-                Action = nameof(Edit)
-            };
-            return View("Forms/CategoryFormPartial", viewModel);
+            return RedirectToAction(nameof(Index), "Home");
         }
 
         [HttpPost]
@@ -74,21 +45,21 @@ namespace PizzeriaButikenOnline.Controllers
             return RedirectToAction(nameof(Index), "Home");
         }
 
+        [HttpDelete("api/categories/{id:int}")]
         public ActionResult Delete(int id)
         {
-            try
-            {
-                var category = _context.Categories.Find(id);
+            var category = _context.Categories.SingleOrDefault(c => c.Id == id);
 
-                _context.Categories.Remove(category);
-                _context.SaveChanges();
+            if (category == null)
+                return NotFound();
 
-                return RedirectToAction(nameof(Index), "Home");
-            }
-            catch
-            {
-                return RedirectToAction(nameof(Index), "Home");
-            }
+            _context.Dishes.RemoveRange(_context.Dishes.Where(d => d.CategoryId == id));
+            _context.SaveChanges();
+
+            _context.Categories.Remove(category);
+            _context.SaveChanges();
+
+            return Ok();
         }
     }
 }
